refactor: extract navigation (goToPage, page nav) into js/navigation.js
